requirements.txt
python-dotenv==1.0.0
pandas==2.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27 